--- restapi/api/api_test.go
@@ -1,9 +1,0 @@
-package api_test
-
-import (
-	"testing"
-)
-
-func TestHealthEndpoint(t *testing.T) {
-
-}
