+++ restapi/api/util.go
@@ -1,6 +1,9 @@
 package api
 
-import "io/ioutil"
+import (
+	"encoding/json"
+	"io/ioutil"
+)
 
 //GetVersion reads the contents of the file version.txt and returns it.
 //If the file cannot be read, it returns "could not read version"
@@ -10,4 +13,12 @@ func GetVersion() string {
 		return "could not read version"
 	}
 	return string(version)
+}
+
+func MustMarshal(v interface{}) string {
+	b, err := json.Marshal(v)
+	if err != nil {
+		panic(err)
+	}
+	return string(b)
 }
